+++ managed_components/espressif__usb_host_msc/src/msc_host.c
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2015-2024 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2015-2026 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -23,7 +23,10 @@
 #include "msc_scsi_bot.h"
 #include "usb/usb_types_ch9.h"
 #include "usb/usb_helpers.h"
+
+#if !CONFIG_IDF_TARGET_LINUX
 #include "soc/soc_memory_layout.h"
+#endif
 
 // MSC driver spin lock
 static portMUX_TYPE msc_lock = portMUX_INITIALIZER_UNLOCKED;
@@ -326,7 +329,7 @@ static bool is_mass_storage_device(uint8_t dev_addr)
     return is_msc_device;
 }
 
-esp_err_t msc_host_handle_events(uint32_t timeout)
+esp_err_t msc_host_handle_events(TickType_t timeout)
 {
     MSC_RETURN_ON_FALSE(s_msc_driver != NULL, ESP_ERR_INVALID_STATE);
 
@@ -375,7 +378,9 @@ static msc_device_t *find_msc_device(usb_device_handle_t device_handle)
 
 static void client_event_cb(const usb_host_client_event_msg_t *event, void *arg)
 {
-    if (event->event == USB_HOST_CLIENT_EVENT_NEW_DEV) {
+    switch (event->event) {
+    case USB_HOST_CLIENT_EVENT_NEW_DEV:
+        ESP_LOGD(TAG, "New device connected");
         if (is_mass_storage_device(event->new_dev.address)) {
             const msc_host_event_t msc_event = {
                 .event = MSC_DEVICE_CONNECTED,
@@ -383,15 +388,47 @@ static void client_event_cb(const usb_host_client_event_msg_t *event, void *arg)
             };
             s_msc_driver->user_cb(&msc_event, s_msc_driver->user_arg);
         }
-    } else if (event->event == USB_HOST_CLIENT_EVENT_DEV_GONE) {
-        msc_device_t *msc_device = find_msc_device(event->dev_gone.dev_hdl);
-        if (msc_device) {
+        break;
+    case USB_HOST_CLIENT_EVENT_DEV_GONE:
+        ESP_LOGD(TAG, "Device suddenly disconnected");
+        msc_device_t *msc_device_gone = find_msc_device(event->dev_gone.dev_hdl);
+        if (msc_device_gone) {
             const msc_host_event_t msc_event = {
                 .event = MSC_DEVICE_DISCONNECTED,
-                .device.handle = msc_device,
+                .device.handle = msc_device_gone,
             };
             s_msc_driver->user_cb(&msc_event, s_msc_driver->user_arg);
         }
+        break;
+
+#ifdef MSC_HOST_SUSPEND_RESUME_API_SUPPORTED
+    case USB_HOST_CLIENT_EVENT_DEV_SUSPENDED:
+        ESP_LOGD(TAG, "Device suspended");
+        msc_device_t *msc_device_susp = find_msc_device(event->dev_suspend_resume.dev_hdl);
+        if (msc_device_susp) {
+            const msc_host_event_t msc_event = {
+                .event = MSC_DEVICE_SUSPENDED,
+                .device.handle = msc_device_susp,
+            };
+            s_msc_driver->user_cb(&msc_event, s_msc_driver->user_arg);
+        }
+        break;
+    case USB_HOST_CLIENT_EVENT_DEV_RESUMED:
+        ESP_LOGD(TAG, "Device resumed");
+        msc_device_t *msc_device_res = find_msc_device(event->dev_suspend_resume.dev_hdl);
+        if (msc_device_res) {
+            const msc_host_event_t msc_event = {
+                .event = MSC_DEVICE_RESUMED,
+                .device.handle = msc_device_res,
+            };
+            s_msc_driver->user_cb(&msc_event, s_msc_driver->user_arg);
+        }
+        break;
+#endif // MSC_HOST_SUSPEND_RESUME_API_SUPPORTED
+
+    default:
+        ESP_LOGW(TAG, "Unrecognized USB Host client event");
+        break;
     }
 }
 
@@ -653,9 +690,13 @@ esp_err_t msc_bulk_transfer(msc_device_t *device, uint8_t *data, size_t size, ms
     switch (status) {
     case USB_TRANSFER_STATUS_COMPLETED:
         if (ep == MSC_EP_IN) {
-            memcpy(data, xfer->data_buffer, xfer->actual_num_bytes);
+            if (xfer->actual_num_bytes > size) {
+                ret = ESP_ERR_INVALID_SIZE;
+            } else {
+                memcpy(data, xfer->data_buffer, xfer->actual_num_bytes);
+                ret = ESP_OK;
+            }
         }
-        ret = ESP_OK;
         break;
     case USB_TRANSFER_STATUS_STALL:
         ret = ESP_ERR_MSC_STALL; break;