+++ managed_components/espressif__usb_host_msc/test_app/pytest_usb_host_msc.py
@@ -1,4 +1,4 @@
-# SPDX-FileCopyrightText: 2024 Espressif Systems (Shanghai) CO LTD
+# SPDX-FileCopyrightText: 2024-2026 Espressif Systems (Shanghai) CO LTD
 # SPDX-License-Identifier: Apache-2.0
 
 from typing import Tuple
@@ -7,13 +7,17 @@ import pytest
 from pytest_embedded_idf.dut import IdfDut
 
 
-@pytest.mark.esp32s2
-@pytest.mark.esp32s3
-@pytest.mark.esp32p4
 @pytest.mark.usb_host
-@pytest.mark.parametrize('count', [
-    2,
-], indirect=True)
+@pytest.mark.parametrize(
+    'count, config, target',
+    [
+        pytest.param(2, 'default', 'esp32s2'),
+        pytest.param(2, 'default', 'esp32s3'),
+        pytest.param(2, 'default', 'esp32p4', marks=[pytest.mark.eco_default]),
+        pytest.param(2, 'esp32p4_eco4', 'esp32p4', marks=[pytest.mark.esp32p4_eco4]),
+    ],
+    indirect=True,
+)
 def test_usb_host_msc(dut: Tuple[IdfDut, IdfDut]) -> None:
     device = dut[0]
     host = dut[1]