+++ managed_components/espressif__usb/CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to this component will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2026-06-16
+
+### Changed
+
+- This component now requires ESP-IDF version >= 5.5.3
+
+### Added
+
+- Automatic root-port suspend before light sleep when `esp_light_sleep_start()` is called (https://github.com/espressif/esp-usb/pull/445), configurable in menuconfig via `CONFIG_ESP_SLEEP_EVENT_CALLBACKS`
+- Added `USB_HOST_CLIENT_EVENT_DEV_REMOVED` and the `notify_dev_removed` client flag so clients can monitor removal of devices they have not opened.
+
+## [1.4.1] - 2026-05-29
+
+### Fixed
+
+- Fixed race between root port disconnect handling and device recycle path that could abort with `ESP_ERR_NOT_FOUND` from `dev_tree_node_dev_gone()` (https://github.com/espressif/esp-idf/issues/18366)
+- List also suspended devices in `usb_host_device_addr_list_fill()` function
+
+## [1.4.0] - 2026-04-02
+
+### Added
+
+- Add ESP32-S31 to supported targets
+
+### Fixed
+
+- Fixed possible out-of-bounds read in `usb_parse_next_descriptor()` in case of malicious Configuration descriptor
+
 ## [1.3.0] - 2026-03-04
 
 ### Added